change abstract to sklearn api
